Final project dog classification/source_transfer_pytorch/requirements.txt
pandas
numpy
nltk
beautifulsoup4
html5lib
Pillow==5.4.1